Refactor provider normalization & scraper fixes
Consolidate provider helpers and clean up scraper/evaluator/dedupe logic. Key changes:

- Remove the legacy "both" LinkedIn+Indeed shortcut from docs, workflow inputs and SOURCE_ALIASES; update README and tests.
- Move several provider utilities into providers.normalization (populated_metadata_dict, seconds_from_published_at, parse_salary_number, format_money, append_metadata_block) and update Indeed/Stepstone to reuse them, simplifying metadata handling.
- Dedupe/merge fixes: use SOURCE_ORDER_INDEX for stable ordering, fix best_text/best_description selection to preserve deterministic tie-breaking, use min() for best_url, and avoid costly lookups in merge/merge helpers.
- Improve matching and index lookups (blocking_index.get default changed to empty tuple) and small scoring/location_similarity parenthesis cleanup.
- Normalization tweaks: precompile whitespace regex, use removeprefix for host normalization, reuse computed description in normalize_job, and simplify description extraction.
- EnvSettings: centralize invalid-value warnings into _warn_invalid and minor read-loop variable rename for clarity.
- Evaluator parsing/storage: cache normalized AI output column names, skip AI columns efficiently, batch Google updates using GOOGLE_BATCH_REQUEST_CHUNK_SIZE and use list.extend for building requests; reorder imports.
- Pipeline dependency check now uses importlib.find_spec to detect missing openai package.
- Scraper/scrape normalization: safer sheet-safe checks, canonical job URL construction uses a local source variable, posted/get_posted behavior simplified to avoid unreliable parsing, format_duration uses rounded seconds.
- Misc: clean up exports in apify client module, minor test no-qa removals and an added test asserting the removed "both" alias.

These changes are primarily refactors and bug fixes to improve determinism, reduce duplication, and centralize provider logic.

+++ src/jobfinder/scraper/service.py
@@ -79,7 +79,7 @@ def parse_output_mode(settings: ScraperSettings) -> set[str]:
 
 def format_duration(seconds: float) -> str:
     """Format elapsed seconds as a compact human-readable duration."""
-    total_seconds = int(round(seconds))
+    total_seconds = round(seconds)
     hours, remainder = divmod(total_seconds, 3600)
     minutes, seconds = divmod(remainder, 60)
